Currently, the program does not build. Continued the major refactoring process—almost ready for testing.

+++ compiler/dataStructures/linkedList/linkedlist.c
@@ -1,10 +1,8 @@
 #include<stdlib.h>
-#include<stdio.h>
-
 #include "linkedlist.h"
 
 
-int add(LinkedList* list, void *data, size_t size)
+int add(LinkedList* list, void *data, int size)
 {
     struct node* addition = malloc(sizeof(*addition));
     addition -> data = malloc(size);
